update: 修改函数命名方式：参数换行

+++ XDGitClient/XDSupport/BaseController/TableViewController/XDPageViewViewController.h
@@ -10,9 +10,9 @@
 
 @interface XDPageViewViewController : XDTableViewController
 
-@property (nonatomic) NSInteger page;//起始页为1
+@property (nonatomic) int page;//起始页为1
 @property (nonatomic) BOOL haveNextPage;
 
-- (void)fetchDataAtPage:(NSInteger)page isHeaderRefresh:(BOOL)isHeaderRefresh;
+- (void)fetchDataAtPage:(int)page isHeaderRefresh:(BOOL)isHeaderRefresh;
 
 @end
