+++ XDGitClient/XDSupport/BaseController/TableViewController/XDPageViewViewController.m
@@ -73,7 +73,7 @@
 
 #pragma mark - public fetch data
 
-- (void)fetchDataAtPage:(NSInteger)page isHeaderRefresh:(BOOL)isHeaderRefresh
+- (void)fetchDataAtPage:(int)page isHeaderRefresh:(BOOL)isHeaderRefresh
 {
     
 }
